+++ superdesk/io/feeding_services/email.py
@@ -13,7 +13,7 @@ import imaplib
 
 from flask import current_app as app
 from superdesk.errors import IngestEmailError
-from superdesk.io.registry import register_feeding_service
+from superdesk.io.registry import register_feeding_service, register_feeding_service_parser
 from superdesk.io.feeding_services import FeedingService
 from superdesk.upload import url_for_media
 
@@ -65,8 +65,6 @@ class EmailFeedingService(FeedingService):
         }
     ]
 
-    parser_restricted_values = ['email_rfc822']
-
     def _test(self, provider):
         self._update(provider, update=None, test=True)
 
@@ -81,7 +79,7 @@ class EmailFeedingService(FeedingService):
                 socket.setdefaulttimeout(app.config.get('EMAIL_TIMEOUT', 10))
                 imap = imaplib.IMAP4_SSL(host=server, port=port)
             except (socket.gaierror, OSError) as e:
-                raise IngestEmailError.emailHostError(exception=e)
+                raise IngestEmailError.emailHostError(exception=e, provider=provider)
 
             try:
                 imap.login(config.get('user', None), config.get('password', None))
@@ -120,3 +118,4 @@ class EmailFeedingService(FeedingService):
 
 
 register_feeding_service(EmailFeedingService)
+register_feeding_service_parser(EmailFeedingService.NAME, 'email_rfc822')